requirements.txt
python-pptx
python-dotenv